api/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
google-generativeai
pydantic